move optimizers to eval/instruction directory
