+++ eval/src/vespa/eval/instruction/dense_pow_as_map_optimizer.cpp
