+++ eval/src/vespa/eval/instruction/dense_remove_dimension_optimizer.h
